+++ src/main/java/org/prgms/voucherProgram/domain/voucher/PercentDiscountVoucher.java
@@ -1,4 +1,4 @@
-package org.prgms.voucherProgram.entity.voucher;
+package org.prgms.voucherProgram.domain.voucher;
 
 import java.util.UUID;
