+++ src/main/java/org/prgms/voucherProgram/domain/voucher/Voucher.java
@@ -1,4 +1,4 @@
-package org.prgms.voucherProgram.entity.voucher;
+package org.prgms.voucherProgram.domain.voucher;
 
 import java.io.Serializable;
 import java.util.UUID;
