+++ src/main/java/org/prgms/voucherProgram/repository/voucher/VoucherRepository.java
@@ -2,7 +2,7 @@ package org.prgms.voucherProgram.repository.voucher;
 
 import java.util.List;
 
-import org.prgms.voucherProgram.entity.voucher.Voucher;
+import org.prgms.voucherProgram.domain.voucher.Voucher;
 
 public interface VoucherRepository {
     Voucher save(Voucher voucher);
