+++ src/main/java/org/prgms/voucherProgram/view/InputView.java
@@ -1,6 +1,6 @@
 package org.prgms.voucherProgram.view;
 
-import org.prgms.voucherProgram.entity.voucher.VoucherType;
+import org.prgms.voucherProgram.domain.voucher.VoucherType;
 
 public interface InputView {
